# services/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anucreative API",
    "description": "CV management backend",
    "version": "0.1.0"
  },
  "paths": {
    "/cv/": {
      "post": {
        "tags": [
          "cv"
        ],
        "summary": "Create Cv",
        "description": "Create a new CV",
        "operationId": "createCv",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CVCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CVResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cv/{slug}": {
      "get": {
        "tags": [
          "cv"
        ],
        "summary": "Get Cv",
        "description": "Get a specific CV by slug, fallback to 'base' if not found",
        "operationId": "getCvBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CVResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "cv"
        ],
        "summary": "Update Cv",
        "description": "Update a CV (partial update)",
        "operationId": "updateCvBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CVUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CVResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cv"
        ],
        "summary": "Delete Cv",
        "description": "Delete a CV",
        "operationId": "deleteCvBySlug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "description": "Health check endpoint",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Award": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "awarder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awarder"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "Award"
      },
      "Basics": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "image": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "future": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Future"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "profiles": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Profile"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profiles"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Basics"
      },
      "CV": {
        "properties": {
          "basics": {
            "$ref": "#/components/schemas/Basics"
          },
          "work": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Work"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work"
          },
          "volunteer": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Volunteer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volunteer"
          },
          "education": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Education"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Education"
          },
          "awards": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Award"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards"
          },
          "certificates": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Certificate"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificates"
          },
          "publications": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Publication"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publications"
          },
          "skills": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Skill"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skills"
          },
          "languages": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Language"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Languages"
          },
          "interests": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Interest"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interests"
          },
          "references": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Reference"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "References"
          },
          "projects": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Project"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projects"
          }
        },
        "type": "object",
        "required": [
          "basics"
        ],
        "title": "CV",
        "description": "Full CV matching JSON CV schema"
      },
      "CVCreate": {
        "properties": {
          "type": {
            "type": "string",
            "pattern": "^(base|role|job|company)$",
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "content": {
            "$ref": "#/components/schemas/CV"
          }
        },
        "type": "object",
        "required": [
          "type",
          "name",
          "slug",
          "content"
        ],
        "title": "CVCreate",
        "description": "Request to create a new CV"
      },
      "CVResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "content": {
            "$ref": "#/components/schemas/CV"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "name",
          "slug",
          "content",
          "created_at",
          "updated_at"
        ],
        "title": "CVResponse",
        "description": "CV response with metadata"
      },
      "CVUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "content": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CV"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CVUpdate",
        "description": "Request to update a CV"
      },
      "Certificate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Certificate"
      },
      "Education": {
        "properties": {
          "institution": {
            "type": "string",
            "title": "Institution"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "studyType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Studytype"
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "score": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "courses": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Courses"
          }
        },
        "type": "object",
        "required": [
          "institution"
        ],
        "title": "Education"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Interest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Interest"
      },
      "Language": {
        "properties": {
          "language": {
            "type": "string",
            "title": "Language"
          },
          "fluency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fluency"
          }
        },
        "type": "object",
        "required": [
          "language"
        ],
        "title": "Language"
      },
      "Location": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "postalCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postalcode"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "countryCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Countrycode"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          }
        },
        "type": "object",
        "title": "Location"
      },
      "Profile": {
        "properties": {
          "network": {
            "type": "string",
            "title": "Network"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "network"
        ],
        "title": "Profile"
      },
      "Project": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "highlights": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlights"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Project"
      },
      "Publication": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "releaseDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Releasedate"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Publication"
      },
      "Reference": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Reference"
      },
      "Skill": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Skill"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "Volunteer": {
        "properties": {
          "organization": {
            "type": "string",
            "title": "Organization"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "highlights": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlights"
          }
        },
        "type": "object",
        "required": [
          "organization"
        ],
        "title": "Volunteer"
      },
      "Work": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "startDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startdate"
          },
          "endDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enddate"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "highlights": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Highlights"
          }
        },
        "type": "object",
        "title": "Work"
      }
    }
  }
}